server/requirements.txt
alembic==1.8.0
aniso8601==9.0.1
blinker==1.4
click==8.1.3
colorama==0.4.4
Flask==2.1.2
Flask-Alembic==2.0.1
Flask-Login==0.6.1
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.11.4
importlib-resources==5.7.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
marshmallow==3.16.0
packaging==21.3
passlib==1.7.4
PyJWT==2.4.0
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.37
Werkzeug==2.1.2
zipp==3.8.0